refactor: move some shared utils to background utils
