+++ src/shared/helpers/time.ts
@@ -117,63 +117,6 @@ export function debounceSync<T extends unknown[], R>(
     });
 }
 
-export class Timeout {
-  private timeout: ReturnType<typeof setTimeout> | null = null;
-  #isPaused = false;
-  #remaining = 0;
-  #startTime = 0;
-
-  constructor(
-    private ms: number,
-    private callback: () => void,
-  ) {
-    if (ms > 0) this.reset(ms);
-  }
-
-  reset(ms: number) {
-    this.clear();
-    this.ms = ms;
-    this.#isPaused = false;
-    this.#startTime = Date.now();
-    this.timeout = setTimeout(this.callback, ms);
-  }
-
-  pause() {
-    if (this.#isPaused) return;
-    this.clear();
-    this.#isPaused = true;
-    this.#remaining = this.ms - (Date.now() - this.#startTime);
-  }
-
-  resume() {
-    if (!this.#isPaused) {
-      throw new Error('Unexpected: Timeout was not paused, cannot resume');
-    }
-    if (this.#remaining > 0) {
-      this.timeout = setTimeout(() => {
-        this.callback();
-        this.reset(this.ms);
-      }, this.#remaining);
-    } else {
-      this.reset(this.ms);
-    }
-  }
-
-  clear() {
-    if (this.timeout !== null) {
-      clearTimeout(this.timeout);
-      this.timeout = null;
-    }
-  }
-}
-
-/**
- * Check if `err` (reason) is result of `AbortSignal.timeout()`
- */
-export const isAbortSignalTimeout = (err: unknown): err is DOMException => {
-  return err instanceof DOMException && err.name === 'TimeoutError';
-};
-
 const REPEATING_INTERVAL_REGEX =
   /^R(\d*)\/(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z)\/(P.+)$/;
